requirements.txt
Django
mysqlclient
social-auth-app-django==2.1.0
cerberus
lxml
cssselect
natural
dateparser
django-automated-logging
git+https://github.com/smiley/steamapi
django-static-precompiler
argon2-cffi
django-redis-cache
redis
python-valve
pypugjs
steam
django-extensions
xmltodict
ruamel.yaml
simplejson
hashids
requests
Faker
rstr
